obj indicators: fix bridges in instances
This changes the saved plane of marked objects to be the disk plane they
are on.

fromLocalInstance requires the passed plane to be the disk plane,
however object.getPlane is the scene plane.

This was causing marked objects to be saved at the location of the
under-bridge tile, instead of the tile they are actually on.