spec counter: add defence drain percentage
This adds an infobox showing what percentage an opponent's defence has
been drained, and broadcasts 0 hit specs.

A config option was added to draw the spec misses, which are drawn with
a blue hitsplat background to visually distinguish them from normal
hits.

Support for the special cases for Tekton with bgs/dwh/em was added.

Co-authored-by: 1Defence <>
Co-authored-by: JaayJee <jacobgoulden@gmail.com>