menu swapper: fix crash performing Use swap on items with submenus
This was calling swap("use") on a submenu menu entry, findIndex() was
finding the "Use" from the top level menu, calling swap() which then was
then passing the submenu MenuEntry to client.setMenuEntries()

optionIndexes used to be heavily used, but now it is only used for built
in swaps and for item "Use" swap. Change it to be lazily computed and be
per-menu.